+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/ga4gh/GA4GHBeaconWSServer.java
@@ -53,7 +53,7 @@ public class GA4GHBeaconWSServer extends EvaWSServer {
     
     @RequestMapping(value = "/beacon", method = RequestMethod.GET)
     public GA4GHBeaconResponse beacon(@RequestParam("referenceName") String chromosome,
-                                      @RequestParam("start") int start,
+                                      @RequestParam("start") long start,
                                       @RequestParam("allele") String allele,
                                       @RequestParam("datasetIds") List<String> studies,
                                       HttpServletResponse response)
